refactor: reset race only after finish
